app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "databricks-sdk>=0.101.0",
    "databricks-sql-connector>=4.2.5",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "sse-starlette>=3.3.3",
    "uvicorn>=0.42.0",
]
